// _wms_reference/stride-wms-app-main/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      containers: {
        Row: {
          container_code: string
          container_type: string | null
          created_at: string
          created_by: string | null
          deleted_at: string | null
          footprint_cu_ft: number | null
          id: string
          is_active: boolean
          location_id: string | null
          status: string
          tenant_id: string
          updated_at: string
          warehouse_id: string | null
        }
        Insert: {
          container_code: string
          container_type?: string | null
          created_at?: string
          created_by?: string | null
          deleted_at?: string | null
          footprint_cu_ft?: number | null
          id?: string
          is_active?: boolean
          location_id?: string | null
          status?: string
          tenant_id: string
          updated_at?: string
          warehouse_id?: string | null
        }
        Update: {
          container_code?: string
          container_type?: string | null
          created_at?: string
          created_by?: string | null
          deleted_at?: string | null
          footprint_cu_ft?: number | null
          id?: string
          is_active?: boolean
          location_id?: string | null
          status?: string
          tenant_id?: string
          updated_at?: string
          warehouse_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "containers_location_id_fkey"
            columns: ["location_id"]
            isOneToOne: false
            referencedRelation: "locations"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "containers_tenant_id_fkey"
            columns: ["tenant_id"]
            isOneToOne: false
            referencedRelation: "tenants"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "containers_warehouse_id_fkey"
            columns: ["warehouse_id"]
            isOneToOne: false
            referencedRelation: "warehouses"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}